+++ includes/player.h
@@ -6,7 +6,7 @@
 /*   By: siferrar <siferrar@student.le-101.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/23 18:52:51 by siferrar          #+#    #+#             */
-/*   Updated: 2020/02/26 06:56:38 by siferrar         ###   ########lyon.fr   */
+/*   Updated: 2020/02/27 08:23:26 by siferrar         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 typedef struct	s_camera
 {
-	double		fov;
+	float		fov;
 	t_fpoint	proj_size;
-	double		proj_dist;
+	float		proj_dist;
 }				t_camera;
 
 typedef struct	s_player
@@ -27,19 +27,19 @@ typedef struct	s_player
 	int			inited;
 	int			as_move;
 	t_fpoint	*pos;
-	double		z;
+	float		z;
 	t_fpoint	*r_pos;
 	t_camera	*cam;
-	double		angle;
+	float		angle;
 	t_fpoint	*step;
-	double		rot_speed;
-	double		speed;
+	float		rot_speed;
+	float		speed;
 	t_ctx		*ctx;
 	void		*brain;
-	void		(*rot)(struct s_player *, double);
+	void		(*rot)(struct s_player *, float);
 	void		(*move)(struct s_player *, int);
 	void		(*sidemove)(struct s_player *, int);
-	void		(*jump)(struct s_player *, double);
+	void		(*jump)(struct s_player *, float);
 	void		(*draw)(struct s_player *, t_ctx *);
 }				t_player;
