+++ srcs/wall_detect.c
@@ -6,7 +6,7 @@
 /*   By: siferrar <siferrar@student.le-101.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/30 22:11:09 by siferrar          #+#    #+#             */
-/*   Updated: 2020/02/26 09:37:21 by siferrar         ###   ########lyon.fr   */
+/*   Updated: 2020/02/27 08:23:26 by siferrar         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ t_fpoint to_point(t_fpoint *p)
 	return (ret);
 }
 
-t_fpoint	closest_grid_h(t_fpoint *p, t_map *m, double angle)
+t_fpoint	closest_grid_h(t_fpoint *p, t_map *m, float angle)
 {
 	t_fpoint closest;
 	
@@ -44,19 +44,19 @@ t_fpoint	closest_grid_h(t_fpoint *p, t_map *m, double angle)
 	return (closest);
 }
 
-t_fpoint	 closest_grid_v(t_fpoint *p, t_map *m, double angle)
+t_fpoint	 closest_grid_v(t_fpoint *p, t_map *m, float angle)
 {
 	t_fpoint closest;
 	
 	if (angle > PI / 2.0 && angle < 2 * PI * 0.75)
-		closest.x = floor(p->x / (double)m->bloc_size) * m->bloc_size;
+		closest.x = floor(p->x / (float)m->bloc_size) * m->bloc_size;
 	else
-		closest.x = floor(p->x / (double)m->bloc_size) * m->bloc_size + m->bloc_size;
+		closest.x = floor(p->x / (float)m->bloc_size) * m->bloc_size + m->bloc_size;
 	closest.y = p->y - (p->x - closest.x) * tan(angle);
 	return (closest);
 }
 
-t_fpoint closest_wall_h(t_brain *b, t_fpoint *p, double angle)
+t_fpoint closest_wall_h(t_brain *b, t_fpoint *p, float angle)
 {
 	t_fpoint offset;
 	t_fpoint cur_point;
@@ -67,12 +67,12 @@ t_fpoint closest_wall_h(t_brain *b, t_fpoint *p, double angle)
 	if (angle > PI && angle < 2 * PI)
 	{
 		offset.y = -(b->map->bloc_size);
-		offset.x = ((double)b->map->bloc_size / tan(angle)) * -1;
+		offset.x = ((float)b->map->bloc_size / tan(angle)) * -1;
 	}
 	else
 	{
 		offset.y = b->map->bloc_size;
-		offset.x = ((double)b->map->bloc_size / tan(angle)) * 1;
+		offset.x = ((float)b->map->bloc_size / tan(angle)) * 1;
 	}
 		//dprintf(1, "H :Offset x: %f -  y: %f\n", offset.x, offset.y);
 	while (is_wall == 0 && cur_point.x < b->map->px_width && cur_point.x > 0)
@@ -93,7 +93,7 @@ t_fpoint closest_wall_h(t_brain *b, t_fpoint *p, double angle)
 	return (cur_point);
 }
 
-t_fpoint closest_wall_v(t_brain *b, t_fpoint *p, double angle)
+t_fpoint closest_wall_v(t_brain *b, t_fpoint *p, float angle)
 {
 	t_fpoint cur_point;
 	t_fpoint offset;
@@ -104,10 +104,10 @@ t_fpoint closest_wall_v(t_brain *b, t_fpoint *p, double angle)
 	if (angle < 2 * PI * 0.75 && angle > PI / 2)
 	{
 		offset.x = -b->map->bloc_size;
-		offset.y = ((double)b->map->bloc_size * tan(angle)) * -1;
+		offset.y = ((float)b->map->bloc_size * tan(angle)) * -1;
 	}else{
 		offset.x = b->map->bloc_size;
-		offset.y = ((double)b->map->bloc_size * tan(angle));
+		offset.y = ((float)b->map->bloc_size * tan(angle));
 	}
 		//dprintf(1, "V: Offset x: %f -  y: %f\n", offset.x, offset.y);
 	while (is_wall == 0 && cur_point.y < b->map->px_height && cur_point.y > 0)
@@ -128,7 +128,7 @@ t_fpoint closest_wall_v(t_brain *b, t_fpoint *p, double angle)
 	return (cur_point);
 }
 
-char		get_wall_side(double angle, int closest)
+char		get_wall_side(float angle, int closest)
 {
 	if (angle >= 0 & angle < PI)
 	{
@@ -150,7 +150,7 @@ char		get_wall_side(double angle, int closest)
 	return ('z');
 }
 
-double to_360(double angle)
+float to_360(float angle)
 {
 	if (angle == 0)
 		angle = 0.1;
@@ -161,13 +161,13 @@ double to_360(double angle)
 	return (angle);
 }
 
-t_detect	dist_to_wall(t_brain *b, t_fpoint *p, double angle)
+t_detect	dist_to_wall(t_brain *b, t_fpoint *p, float angle)
 {
 	t_fpoint closest_h;
 	t_fpoint closest_v;
 	t_fpoint dists;
 	t_detect wall;
-	double bad_dist;
+	float bad_dist;
 
 	angle = to_360(angle);
 	closest_h = closest_wall_h(b, p, angle);
@@ -207,12 +207,12 @@ t_buff		**get_wall_texture(t_map *m, char w_side)
 
 void	draw_walls(t_brain *b, t_ctx *c)
 {
-	double w_height;
+	float w_height;
 	t_detect wall;
-	double dist;
-	double cur_col;
-	double col_step;
-	double cur_angle;
+	float dist;
+	float cur_col;
+	float col_step;
+	float cur_angle;
 
 	cur_col = 0;
 	col_step = b->player->cam->fov/ c->width;
@@ -238,12 +238,12 @@ void	draw_walls(t_brain *b, t_ctx *c)
 	}
 }
 /*
-void draw_floor(t_brain *b, double w_height, double cur_col, t_detect w)
+void draw_floor(t_brain *b, float w_height, float cur_col, t_detect w)
 {
 	
 }
 */
-void draw_col(t_brain *b, double w_height, double cur_col, t_detect w)
+void draw_col(t_brain *b, float w_height, float cur_col, t_detect w)
 {
 	int i;
 	t_fpoint ratio;
